@@ -161,17 +167,17 @@
     "karma-jasmine": "~4.0.1",
     "karma-jasmine-html-reporter": "^1.7.0",
     "ngrx-store-freeze": "^0.2.4",
-    "patch-package": "^6.4.7",
+    "patch-package": "^6.5.1",
     "postinstall-prepare": "^2.0.0",
     "protractor": "~7.0.0",
     "raw-loader": "^4.0.2",
-    "ts-node": "^10.4.0",
+    "ts-node": "^10.9.1",
     "typescript": "~4.6.4",
-    "webpack": "^5.64.4"
+    "webpack": "^5.75.0"
   },
   "resolutions": {
-    "lodash": "~4.17.21",
     "@types/react": "17.0.37",
-    "ace-builds": "1.4.13"
+    "ace-builds": "1.4.13",
+    "@date-io/core": "1.3.7"
   }
 }
