@@ -72,7 +72,7 @@ export class RateLimitsListComponent implements ControlValueAccessor, Validator,
   rateLimitsArray: Array<RateLimits>;
 
   private valueChangeSubscription: Subscription = null;
-  private destroy$ = new Subject();
+  private destroy$ = new Subject<void>();
   private propagateChange = (v: any) => { };
 
   constructor(private fb: UntypedFormBuilder) {}
