@@ -52,7 +52,7 @@ import { Store } from '@ngrx/store';
 import { AppState } from '@core/core.state';
 import { TranslateService } from '@ngx-translate/core';
 import { EntityType, entityTypeTranslations } from '@shared/models/entity-type.models';
-import { MatCheckbox } from '@angular/material/checkbox/checkbox';
+import { MatCheckbox } from '@angular/material/checkbox';
 import { TbPopoverService } from '@shared/components/popover.service';
 import { RemoveOtherEntitiesConfirmComponent } from '@home/components/vc/remove-other-entities-confirm.component';
 import { entityGroupTypes } from '@shared/models/entity-group.models';
@@ -232,7 +232,7 @@ export class EntityTypesVersionLoadComponent extends PageComponent implements On
   allowedEntityTypes(entityTypeControl?: AbstractControl): Array<EntityType> {
     let res = [...exportableEntityTypes];
     const currentEntityType: EntityType = entityTypeControl?.get('entityType')?.value;
-    const value: [{entityType: string, config: EntityTypeVersionLoadConfig}] =
+    const value: [{entityType: string; config: EntityTypeVersionLoadConfig}] =
       this.entityTypesVersionLoadFormGroup.get('entityTypes').value || [];
     const usedEntityTypes = value.map(val => val.entityType).filter(val => val);
     res = res.filter(entityType => !usedEntityTypes.includes(entityType) || entityType === currentEntityType);
@@ -267,7 +267,7 @@ export class EntityTypesVersionLoadComponent extends PageComponent implements On
   }
 
   private updateModel() {
-    const value: [{entityType: string, config: EntityTypeVersionLoadConfig}] =
+    const value: [{entityType: string; config: EntityTypeVersionLoadConfig}] =
       this.entityTypesVersionLoadFormGroup.get('entityTypes').value || [];
     let modelValue: {[entityType: string]: EntityTypeVersionLoadConfig} = null;
     if (value && value.length) {
