@@ -122,7 +122,7 @@ export class WidgetComponentService {
           }, new WidgetTypeId('1'), new TenantId( NULL_UUID ), 'customWidgetBundle', undefined
         );
       }
-      const initSubject = new ReplaySubject();
+      const initSubject = new ReplaySubject<void>();
       this.init$ = initSubject.asObservable();
 
       const w = (this.window as any);
