@@ -40,7 +40,7 @@ import {
   Output,
   ViewChild
 } from '@angular/core';
-import { MatFormFieldAppearance } from '@angular/material/form-field/form-field';
+import { MatFormFieldAppearance } from '@angular/material/form-field';
 import { ControlValueAccessor, UntypedFormBuilder, UntypedFormGroup, NG_VALUE_ACCESSOR } from '@angular/forms';
 import { merge, Observable, of, Subject } from 'rxjs';
 import { catchError, debounceTime, map, share, switchMap, tap } from 'rxjs/operators';
