@@ -103,7 +103,7 @@ export class TbMarkdownComponent implements OnChanges {
   }
 
   private render(markdown: string) {
-    const compiled = this.markdownService.compile(markdown, false);
+    const compiled = this.markdownService.parse(markdown, { decodeHtml: false });
     let template = this.sanitizeCurlyBraces(compiled);
     let markdownClass = 'tb-markdown-view';
     if (this.markdownClass) {
