@@ -45,6 +45,7 @@ import { PopoverPlacement, PopoverWithTrigger } from '@shared/components/popover
 import { TbPopoverComponent } from '@shared/components/popover.component';
 import { ComponentType } from '@angular/cdk/portal';
 import { HELP_MARKDOWN_COMPONENT_TOKEN } from '@shared/components/tokens';
+import { CdkOverlayOrigin } from '@angular/cdk/overlay';
 
 @Injectable()
 export class TbPopoverService {
@@ -99,7 +100,7 @@ export class TbPopoverService {
       componentRef.location.nativeElement
     );
     const originElementRef = new ElementRef(trigger);
-    component.setOverlayOrigin({ elementRef: originElementRef });
+    component.setOverlayOrigin(new CdkOverlayOrigin(originElementRef));
     component.tbPlacement = preferredPlacement;
     component.tbComponentFactory = this.resolver.resolveComponentFactory(componentType);
     component.tbComponentInjector = injector;
@@ -147,7 +148,7 @@ export class TbPopoverService {
       const originElementRef = new ElementRef(trigger);
       component.tbAnimationState = 'void';
       component.tbOverlayStyle = {...overlayStyle, opacity: '0' };
-      component.setOverlayOrigin({ elementRef: originElementRef });
+      component.setOverlayOrigin(new CdkOverlayOrigin(originElementRef));
       component.tbPlacement = preferredPlacement;
       component.tbComponentFactory = this.resolver.resolveComponentFactory(this.helpMarkdownComponent);
       component.tbComponentInjector = injector;
